+++ lib/Config.cpp
@@ -230,7 +230,7 @@ Albinos::ReturnedValue Albinos::Config::removeSetting(char const *name)
 ///
 /// \todo implementation
 ///
-Albinos::ReturnedValue Albinos::Config::include(Key *inheritFrom, size_t position)
+Albinos::ReturnedValue Albinos::Config::include(Key *inheritFrom, int position)
 {
   (void)inheritFrom;
   (void)position;
